Updated code and build complete.

+++ main/tuning_http_server.c
@@ -1,5 +1,5 @@
 #include "tuning_http_server.h"
-
+// #include <inttypes.h>
 
 static const char *TAG = "websocket_server";
 
@@ -108,10 +108,10 @@ void websocket_callback(uint8_t num, WEBSOCKET_TYPE_t type, char *msg, uint64_t 
         }
         break;
     case WEBSOCKET_BIN:
-        ESP_LOGI(TAG, "client %i sent binary message of size %i:\n%s", num, (uint32_t)len, msg);
+        ESP_LOGI(TAG, "client %i sent binary message of size %llu:\n%s", num, (unsigned long long)len, msg);
         break;
     case WEBSOCKET_PING:
-        ESP_LOGI(TAG, "client %i pinged us with message of size %i:\n%s", num, (uint32_t)len, msg);
+        ESP_LOGI(TAG, "client %i pinged us with message of size %llu:\n%s", num, (unsigned long long)len, msg);
         break;
     case WEBSOCKET_PONG:
         ESP_LOGI(TAG, "client %i responded to the ping", num);
